requirements.txt
SpeechRecognition
PyAudio
streamlit
google-generativeai
pipwin
python-dotenv

-e .